Create documentation
Create project's documentation using sphinx. The documentation includes
the infomation available in the README.rst file, as well as the autodoc
generated with sphinx-apidoc. Windows make and Linux Makefile where
also included.

A new logo file was created to be used in the documentation, however is
a bit nasty so it should be replaced with a proper one.